# docs/api/openapi.json
{
  "components": {
    "schemas": {
      "AdminHeatmapCell": {
        "properties": {
          "n_a": {
            "default": 0,
            "title": "N A",
            "type": "integer"
          },
          "not_yet": {
            "default": 0,
            "title": "Not Yet",
            "type": "integer"
          },
          "yes": {
            "default": 0,
            "title": "Yes",
            "type": "integer"
          }
        },
        "title": "AdminHeatmapCell",
        "type": "object"
      },
      "AdminHeatmapResponse": {
        "properties": {
          "matrix": {
            "additionalProperties": {
              "additionalProperties": {
                "additionalProperties": {
                  "$ref": "#/components/schemas/AdminHeatmapCell"
                },
                "type": "object"
              },
              "type": "object"
            },
            "title": "Matrix",
            "type": "object"
          },
          "topic_structure": {
            "additionalProperties": {
              "items": {
                "additionalProperties": true,
                "type": "object"
              },
              "type": "array"
            },
            "title": "Topic Structure",
            "type": "object"
          },
          "total_submitted": {
            "title": "Total Submitted",
            "type": "integer"
          }
        },
        "required": [
          "total_submitted",
          "matrix",
          "topic_structure"
        ],
        "title": "AdminHeatmapResponse",
        "type": "object"
      },
      "AdminInitiativeRow": {
        "properties": {
          "answer_count": {
            "title": "Answer Count",
            "type": "integer"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "participant_type": {
            "title": "Participant Type",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "user_email": {
            "title": "User Email",
            "type": "string"
          }
        },
        "required": [
          "id",
          "user_email",
          "name",
          "participant_type",
          "status",
          "created_at",
          "answer_count"
        ],
        "title": "AdminInitiativeRow",
        "type": "object"
      },
      "AdminUserRow": {
        "properties": {
          "answer_count": {
            "default": 0,
            "title": "Answer Count",
            "type": "integer"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "initiative_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Initiative Name"
          },
          "initiative_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Initiative Status"
          },
          "participant_type": {
            "title": "Participant Type",
            "type": "string"
          },
          "role": {
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "role",
          "participant_type",
          "created_at"
        ],
        "title": "AdminUserRow",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "InitiativeCreate": {
        "properties": {
          "contact_email": {
            "anyOf": [
              {
                "format": "email",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Email"
          },
          "contact_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "organization": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Organization"
          },
          "sector": {
            "title": "Sector",
            "type": "string"
          },
          "sector_other": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sector Other"
          }
        },
        "required": [
          "name",
          "sector"
        ],
        "title": "InitiativeCreate",
        "type": "object"
      },
      "InitiativeRead": {
        "properties": {
          "contact_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Email"
          },
          "contact_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Name"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "organization": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Organization"
          },
          "participant_type": {
            "title": "Participant Type",
            "type": "string"
          },
          "sector": {
            "title": "Sector",
            "type": "string"
          },
          "sector_other": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sector Other"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          },
          "user_id": {
            "title": "User Id",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "user_id",
          "name",
          "description",
          "sector",
          "sector_other",
          "contact_name",
          "contact_email",
          "organization",
          "participant_type",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "InitiativeRead",
        "type": "object"
      },
      "InitiativeStatus": {
        "enum": [
          "draft",
          "active",
          "submitted"
        ],
        "title": "InitiativeStatus",
        "type": "string"
      },
      "InitiativeUpdate": {
        "properties": {
          "contact_email": {
            "anyOf": [
              {
                "format": "email",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Email"
          },
          "contact_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "organization": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Organization"
          },
          "sector": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sector"
          },
          "sector_other": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sector Other"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InitiativeStatus"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "InitiativeUpdate",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "contact": {
      "name": "CoE-DSC",
      "url": "https://coe-dsc.nl/"
    },
    "description": "API for the MAMI Framework DSI Assessment Tool \u2014 CoE-DSC / TNO",
    "title": "MAMI Checker API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/admin/export": {
      "get": {
        "description": "Stream all initiatives + answers as a CSV file download.\n\nNote: evidence URLs are stored in a separate EvidenceURL table and are not\nincluded in this export. The CSV contains one row per questionnaire answer.",
        "operationId": "export_dataset_api_v1_admin_export_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Export Dataset",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/heatmap": {
      "get": {
        "description": "Aggregate yes/not_yet/n_a counts per heatmap cell across all submitted initiatives.",
        "operationId": "get_admin_heatmap_api_v1_admin_heatmap_get",
        "parameters": [
          {
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Type"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminHeatmapResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Admin Heatmap",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/initiatives": {
      "get": {
        "description": "List all initiatives with owner email, status, and answer count.",
        "operationId": "list_initiatives_api_v1_admin_initiatives_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AdminInitiativeRow"
                  },
                  "title": "Response List Initiatives Api V1 Admin Initiatives Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "List Initiatives",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/initiatives/{initiative_id}": {
      "delete": {
        "description": "Delete an initiative and its child rows (answers, evidence, report), keeping the user.",
        "operationId": "delete_initiative_api_v1_admin_initiatives__initiative_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "initiative_id",
            "required": true,
            "schema": {
              "title": "Initiative Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Delete Initiative",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/reset-demo": {
      "post": {
        "description": "Delete all non-admin users and all their data. DESTRUCTIVE.\n\nUsed by demo organizer to clean up between demo runs.\nAdmin users are never deleted.",
        "operationId": "reset_demo_api_v1_admin_reset_demo_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Reset Demo",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/users": {
      "get": {
        "description": "List all registered users with their initiative status.",
        "operationId": "list_users_api_v1_admin_users_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AdminUserRow"
                  },
                  "title": "Response List Users Api V1 Admin Users Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "List Users",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/users/{user_id}": {
      "delete": {
        "description": "Hard-delete a user and all their data (initiative, answers, evidence, report).",
        "operationId": "delete_user_api_v1_admin_users__user_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Delete User",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/api/v1/initiatives": {
      "post": {
        "operationId": "create_initiative_api_v1_initiatives_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InitiativeCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InitiativeRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Create Initiative",
        "tags": [
          "initiatives"
        ]
      }
    },
...
    "/api/v1/initiatives/{initiative_id}": {
      "patch": {
        "operationId": "update_initiative_api_v1_initiatives__initiative_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "initiative_id",
            "required": true,
            "schema": {
              "title": "Initiative Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InitiativeUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InitiativeRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Update Initiative",
        "tags": [
          "initiatives"
        ]
      }
    },
...
  }
}